requirements.txt
autopep8==2.0.2
backcall==0.2.0
colorama==0.4.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
fonttools==4.38.0
imageio==2.26.0
importlib-metadata==6.0.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
llvmlite==0.39.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.6.3
numba==0.56.4
numpy==1.21.6
opencv-python==4.7.0.72
packaging==23.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
psutil==5.9.4
pycodestyle==2.10.0
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
pywin32==305
pyzmq==25.0.0
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0.post1
tensorly==0.8.0
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
zipp==3.15.0
